% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{lob_constants_filing_types}
\alias{lob_constants_filing_types}
\alias{lob_constants_lobbying_activity_issues}
\alias{lob_constants_government_entities}
\alias{lob_constants_countries}
\alias{lob_constants_states}
\alias{lob_constants_prefixes}
\alias{lob_constants_suffixes}
\alias{lob_constants_contribution_item_types}
\title{Get constants}
\usage{
lob_constants_filing_types()

lob_constants_lobbying_activity_issues()

lob_constants_government_entities()

lob_constants_countries()

lob_constants_states()

lob_constants_prefixes()

lob_constants_suffixes()

lob_constants_contribution_item_types()
}
\value{
a \code{tibble} of lobbyists
}
\description{
Get constants
}
\examples{
\dontshow{if (has_lobby_key()) withAutoprint(\{ # examplesIf}
lob_constants_filing_types()
lob_constants_lobbying_activity_issues()
lob_constants_government_entities()
lob_constants_countries()
lob_constants_states()
lob_constants_prefixes()
lob_constants_suffixes()
lob_constants_contribution_item_types()
\dontshow{\}) # examplesIf}
}
