% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_lobbyists.R
\name{lob_list_lobbyists}
\alias{lob_list_lobbyists}
\title{List Lobbyists}
\usage{
lob_list_lobbyists(
  id = NULL,
  lobbyist_name = NULL,
  ordering = NULL,
  page = NULL,
  page_size = NULL,
  registrant_id = NULL,
  registrant_name = NULL,
  clean = TRUE
)
}
\arguments{
\item{id}{ID. Integer value.}

\item{lobbyist_name}{Lobbyist Name. String value.}

\item{ordering}{Which field to use when ordering the results. String value.}

\item{page}{A page number within the paginated result set. Integer value.}

\item{page_size}{Number of results to return per page. Integer value.}

\item{registrant_id}{Registrant ID. Integer value.}

\item{registrant_name}{Registrant Name. String value.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \code{tibble} of lobbyists
}
\description{
List Lobbyists
}
\examples{
\dontshow{if (has_lobby_key()) withAutoprint(\{ # examplesIf}
lob_list_lobbyists()
\dontshow{\}) # examplesIf}
}
