% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_next.R
\name{lob_request_next}
\alias{lob_request_next}
\title{Request next set of responses}
\usage{
lob_request_next(response, max_req = 1)
}
\arguments{
\item{response}{A \link[tibble:tibble]{tibble::tibble} from a \verb{lob_*} function}

\item{max_req}{A max number of additional requests to make. Default is 1.}
}
\value{
a tibble with responses bound by row to new results
}
\description{
Request next set of responses
}
\examples{
\dontshow{if (has_lobby_key()) withAutoprint(\{ # examplesIf}
lob_list_filings(filing_year = 2024) |>
  lob_request_next()
\dontshow{\}) # examplesIf}
}
