% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_client.R
\name{lob_retrieve_client}
\alias{lob_retrieve_client}
\title{Retrieve a specific client by its ID}
\usage{
lob_retrieve_client(id, clean = TRUE)
}
\arguments{
\item{id}{ID. Integer value.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \code{tibble} of lobbying client details
}
\description{
Retrieve a specific client by its ID
}
\examples{
\dontshow{if (has_lobby_key()) withAutoprint(\{ # examplesIf}
lob_retrieve_client(id = '44400')
\dontshow{\}) # examplesIf}
}
