% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_contribution.R
\name{lob_retrieve_contribution}
\alias{lob_retrieve_contribution}
\title{Retrieve a specific contribution by its UUID}
\usage{
lob_retrieve_contribution(filing_uuid, clean = TRUE)
}
\arguments{
\item{filing_uuid}{filing_uuid. UUID value.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \code{tibble} of lobbying contribution details
}
\description{
Retrieve a specific contribution by its UUID
}
\examples{
\dontshow{if (has_lobby_key()) withAutoprint(\{ # examplesIf}
lob_retrieve_contribution(filing_uuid = '86260004-84e7-46e3-9cfa-76edae508768')
\dontshow{\}) # examplesIf}
}
