% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_filing.R
\name{lob_retrieve_filing}
\alias{lob_retrieve_filing}
\title{Retrieve a specific filing by its UUID}
\usage{
lob_retrieve_filing(filing_uuid, clean = TRUE)
}
\arguments{
\item{filing_uuid}{filing_uuid. UUID value.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \code{tibble} of lobbying filing details
}
\description{
Retrieve a specific filing by its UUID
}
\examples{
\dontshow{if (has_lobby_key()) withAutoprint(\{ # examplesIf}
lob_retrieve_filing(filing_uuid = '3b181214-b253-462c-ad64-783c827e159e')
\dontshow{\}) # examplesIf}
}
