% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_lobbyist.R
\name{lob_retrieve_lobbyist}
\alias{lob_retrieve_lobbyist}
\title{Retrieve a specific lobbyist by their ID}
\usage{
lob_retrieve_lobbyist(id, clean = TRUE)
}
\arguments{
\item{id}{ID. Integer value.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \code{tibble} of lobbyist
}
\description{
Retrieve a specific lobbyist by their ID
}
\examples{
\dontshow{if (has_lobby_key()) withAutoprint(\{ # examplesIf}
lob_retrieve_lobbyist(id = 1)
\dontshow{\}) # examplesIf}
}
