% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_registrant.R
\name{lob_retrieve_registrant}
\alias{lob_retrieve_registrant}
\title{Retrieve a specific registrant by their ID}
\usage{
lob_retrieve_registrant(id, clean = TRUE)
}
\arguments{
\item{id}{ID. Integer value.}

\item{clean}{Default is TRUE. Should output be returned as a \code{tibble} (\code{TRUE}) or list.}
}
\value{
a \code{tibble} of lobbying registrant details
}
\description{
Retrieve a specific registrant by their ID
}
\examples{
\dontshow{if (has_lobby_key()) withAutoprint(\{ # examplesIf}
lob_retrieve_registrant(id = '11')
\dontshow{\}) # examplesIf}
}
