% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobby-package.R
\docType{package}
\name{lobby-package}
\alias{lobby}
\alias{lobby-package}
\title{lobby: Interact with the 'US Senate Lobbying Disclosure API'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Download and read data on lobbying in the United States Congress. Data is queried from the Senate's Application Programming Interface (\url{https://lda.senate.gov/api/}). This supports filings since 2008. Functions exist for all primary data endpoints, including queries by filings, contributions, registrations, clients, and lobbyists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christopherkenny/lobby}
  \item \url{https://christophertkenny.com/lobby/}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
