% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobby_key.R
\name{set_lobby_key}
\alias{set_lobby_key}
\title{Add Entry to Renviron}
\usage{
set_lobby_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{Character. API key to add to add.}

\item{overwrite}{Defaults to FALSE. Boolean. Should existing \code{USSLDA_KEY} in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Boolean. Should this be added '~/.Renviron' file?}
}
\value{
key, invisibly
}
\description{
Adds US Senate Lobbying Disclosure API key to .Renviron.
}
\examples{
\dontrun{
set_lobby_key('1234')
}

}
