% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lobby_key.R
\name{key}
\alias{key}
\alias{has_lobby_key}
\alias{get_lobby_key}
\title{Check or Get US Senate Lobbying Disclosure API Key}
\usage{
has_lobby_key()

get_lobby_key()
}
\value{
logical if \code{has}, key if \code{get}
}
\description{
Check or Get US Senate Lobbying Disclosure API Key
}
\examples{
has_lobby_key()
}
